etl/requirements.txt
requests
pandas
sqlalchemy
python-dotenv
